+++ src/Bootstrap.php
@@ -14,6 +14,7 @@ declare( strict_types = 1 );
 
 namespace Liebherr\InterfaceWorld;
 
+use Liebherr\InterfaceWorld\Admin\AdminAssets;
 use Liebherr\InterfaceWorld\Admin\AdminMenu;
 use Liebherr\InterfaceWorld\Connection\ConnectionMapView;
 use Liebherr\InterfaceWorld\CoreBridge\MediaBridge;
@@ -39,6 +40,7 @@ final class Bootstrap {
 
 		if ( is_admin() ) {
 			AdminMenu::register();
+			AdminAssets::register();
 		}
 
 		add_action( 'admin_notices', [ self::class, 'maybe_show_permalink_notice' ] );